requirements.txt
#torch>=1.7.0
#torchvision>=0.8.1
pytorch-lightning>=1.0.8
requests>=2.25.0
matplotlib>=3.3.3
Pillow>=8.0.1
scikit-image
opencv-python>=4.4.0.46
scipy>=1.5.4
facenet-pytorch>=2.5.1
torchsummary>=1.5.1
pandas>=1.1.5
scikit-learn>=0.24.0
optuna
unrar